+++ locallib.php
@@ -14,30 +14,31 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-/**
- * Provides the classes used internally in the plugin
- *
- * @package     report_overviewstats
- * @copyright   2013 David Mudrak <david@moodle.com>
- * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
+ /**
+  * Provides the classes used internally in the plugin
+  *
+  * @package report_overviewstats
+  * @author DualCube <admin@dualcube.com>
+  * @copyright 2023 DualCube <admin@dualcube.com>
+  * @copyright based on work by 2013 David Mudrak <david@moodle.com>
+  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+  */
 
 defined('MOODLE_INTERNAL') || die();
 
 // Need to load the base classes first so we can extend them.
-require_once($CFG->dirroot.'/report/overviewstats/classes/chart.php');
-require_once($CFG->dirroot.'/report/overviewstats/classes/manager.php');
+require_once($CFG->dirroot . '/report/overviewstats/classes/chart.php');
 
 // Load all classes files (to be replaced by autoloading in Moodle 2.6.
-$classfiles = new DirectoryIterator($CFG->dirroot.'/report/overviewstats/classes/');
+$classfiles = new DirectoryIterator($CFG->dirroot . '/report/overviewstats/classes/');
 foreach ($classfiles as $classfile) {
     if ($classfile->isDot()) {
         continue;
     }
     if ($classfile->isLink()) {
-        throw new coding_exception('Unexpected symlink in report/overviewstats/classes/');
+        throw new coding_exception(get_string('Unexpected-symlink-exception', 'report_overviewstats'));
     }
-    if ($classfile->isFile() and substr($classfile->getFilename(), -4) === '.php') {
+    if ($classfile->isFile() && substr($classfile->getFilename(), -4) === '.php') {
         require_once($classfile->getPathname());
     }
 }