+++ settings.php
@@ -14,21 +14,22 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-/**
- * Contains settings that the user can customize for the plugin
- * 
- * @package     report_overviewstats
- * @category    admin
- * @copyright   2013 David Mudrak <david@moodle.com>
- * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
+ /**
+  * Contains settings that the user can customize for the plugin
+  *
+  * @package report_overviewstats
+  * @category admin
+  * @author DualCube <admin@dualcube.com>
+  * @copyright 2023 DualCube <admin@dualcube.com>
+  * @copyright based on work by 2013 David Mudrak <david@moodle.com>
+  * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+  */
 
 defined('MOODLE_INTERNAL') || die();
 
 $ADMIN->add('reports', new admin_externalpage('overviewstats',
     get_string('pluginname', 'report_overviewstats'),
-    $CFG->wwwroot.'/report/overviewstats/index.php', 'report/overviewstats:view'));
+    $CFG->wwwroot . '/report/overviewstats/index.php', 'report/overviewstats:view'));
 
 // No report settings.
 $settings = null;
-